[FIX] Bug: 260987 - Unable to insert NULL to timestamp with time zone
The database server would complain
    ('Cannot convert integer to timestamp with time zone', -157)
when a python script was trying to insert Nulls to timestamp with time
zone columns through parameters.  The problem is the driver was trying
to set the data type to A_VAL32 to pass the Null value to the server
and the server does not like it for a timestamp with time zone column
even the data is null.